// packages/holy-grail-api/src/domains/services/providers/cloudflare/interfaces/dns-record.ts
export default interface CF_DNSRecord {
  id: string;
  type: string;
  name: string;
  content: string;
  proxiable: boolean;
  proxied: boolean;
  ttl: number;
  locked: boolean;
  zone_id: string;
  zone_name: string;
  created_on: string;
  modified_on: string;
  data: any;
  meta: {
    auto_added: boolean;
    source: string;
  }
}